Add files via upload
Apply CSS to an existing HTML website

Start a new git project named A1. Make sure to commit and sync your changes after every requirement is complete, The commit message should contain the requirement(s) that was addressed in that commit.

Using the HTML you created in Assignment1,

- Link ONE external CSS stylesheet to the HTML page. Do not use inline or document-wide styles.

- Apply element, id and class selectors where appropriate to apply the following visual updates:
- body font size should be 16px. Use a custom google font of your choice for the font face.
- headings should be 1.5 times the size of the parent container font.
- demonstrate your understanding of the box model by using width, height, margin, border and padding.
- apply a background image to your site that takes up the entire viewport.
- body background colour should be grey
- article background colours should be white.
- featured articles should have a greater font size than other articles
- all "Technical" article authors should have a black background with white text.
- all "Career" article authors should have a purple background with white text.

header and footer must not scroll with the document and must be displayed over all other content.
All headings must be placed 30 units from the left of it's own origin position. (your choice on values and units)

On a mobile device,
- Apply a one column layout for all components so that they flow one under the other.
- Increase the font size and line spacing of paragraphs
- Show a small logo

On desktop and tablet, apply the following layout:
- header and footer must take up the entire width of the page
- The featured article should take up 80% of the page
- Each category should be laid out under the featured article in 3 columns, 1 row
- Articles in each category should flow one under the other.
- Show a big logo

- The borders of all required textfields in the contact form must be styled red.

These styles must be applied to all pages.